+++ packages/button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/button
 
+## 1.0.9
+
+### Patch Changes
+
+- Fix button icon
+- Updated dependencies
+  - @pikas-ui/icons@1.0.9
+  - @pikas-ui/loader@1.0.9
+  - @pikas-ui/styles@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes